Add support for get_client_name() and get_client_id() to engine
mididings.engine.get_client_name() will return the unique name of the mididings
sequencer engine, expecially useful for jack, which uses unique names for
clients and thus will add "-01" (02, 03, ...) to the client_name set in the
mididings.config().

mididings.engine.get_client_id() will return the client port number (ALSA) or
client UUID (JACK), which is useful for ALSA, which instead allows clients with
the same name. The result is an integer for ALSA, while for JACK is usually a
(long) integer, but there's support for strings too (jack returns the UUID as
a string, I actually don't know why).